fampay/requirements.txt
Django
django-cors-headers
djangorestframework
django-extensions
django-filter
google-api-python-client
psycopg2-binary
gunicorn
python-dateutil
ipython
